+++ Steamfitter.Api/Controllers/ScenarioTemplateController.cs
@@ -6,9 +6,9 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using STT = System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Steamfitter.Api.Infrastructure.Extensions;
+using Steamfitter.Api.Data;
+using Steamfitter.Api.Infrastructure.Authorization;
 using Steamfitter.Api.Infrastructure.Exceptions;
 using Steamfitter.Api.Services;
 using SAVM = Steamfitter.Api.ViewModels;
@@ -19,9 +19,9 @@ namespace Steamfitter.Api.Controllers
     public class ScenarioTemplateController : BaseController
     {
         private readonly IScenarioTemplateService _scenarioTemplateService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly ISteamfitterAuthorizationService _authorizationService;
 
-        public ScenarioTemplateController(IScenarioTemplateService scenarioTemplateService, IAuthorizationService authorizationService)
+        public ScenarioTemplateController(IScenarioTemplateService scenarioTemplateService, ISteamfitterAuthorizationService authorizationService)
         {
             _scenarioTemplateService = scenarioTemplateService;
             _authorizationService = authorizationService;
@@ -41,45 +41,12 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "getScenarioTemplates")]
         public async STT.Task<IActionResult> Get(CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.ViewScenarioTemplates], ct))
+                throw new ForbiddenException();
+
             var list = await _scenarioTemplateService.GetAsync(ct);
             return Ok(list);
         }
-
-        // /// <summary>
-        // /// Gets all ScenarioTemplates for a User
-        // /// </summary>
-        // /// <remarks>
-        // /// Returns all ScenarioTemplates where the specified User is a member of at least one of it's Teams
-        // /// <para />
-        // /// Accessible to a SuperUser or the specified User itself
-        // /// </remarks>
-        // /// <returns></returns>
-        // [HttpGet("users/{id}/scenarioTemplates")]
-        // [ProducesResponseType(typeof(IEnumerable<ScenarioTemplate>), (int)HttpStatusCode.OK)]
-        // [SwaggerOperation(OperationId = "getUserScenarioTemplates")]
-        // public async STT.Task<IActionResult> GetByUserId(Guid id, CancellationToken ct)
-        // {
-        //     var list = await _scenarioTemplateService.GetByUserIdAsync(id, ct);
-        //     return Ok(list);
-        // }
-
-        // /// <summary>
-        // /// Gets all ScenarioTemplates for the current User
-        // /// </summary>
-        // /// <remarks>
-        // /// Returns all ScenarioTemplates where the current User is a member of at least one of it's Teams
-        // /// <para />
-        // /// Accessible only to the current User.
-        // /// <para/>
-        // /// This is a convenience endpoint and simply returns a 302 redirect to the fully qualified users/{id}/scenarioTemplates endpoint
-        // /// </remarks>
-        // [HttpGet("me/scenarioTemplates")]
-        // [ProducesResponseType(typeof(IEnumerable<ScenarioTemplate>), (int)HttpStatusCode.OK)]
-        // [SwaggerOperation(OperationId = "getMyScenarioTemplates")]
-        // public async STT.Task<IActionResult> GetMy(CancellationToken ct)
-        // {
-        //     return RedirectToAction(nameof(this.GetByUserId), new { id = User.GetId() });
-        // }
 
         /// <summary>
         /// Gets a specific ScenarioTemplate by id
@@ -97,10 +64,10 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "getScenarioTemplate")]
         public async STT.Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
-            var scenarioTemplate = await _scenarioTemplateService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ViewScenarioTemplates], [ScenarioTemplatePermission.ViewScenarioTemplate], ct))
+                throw new ForbiddenException();
 
-            if (scenarioTemplate == null)
-                throw new EntityNotFoundException<SAVM.ScenarioTemplate>();
+            var scenarioTemplate = await _scenarioTemplateService.GetAsync(id, ct);
 
             return Ok(scenarioTemplate);
         }
@@ -120,6 +87,9 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "createScenarioTemplate")]
         public async STT.Task<IActionResult> Create([FromBody] SAVM.ScenarioTemplateForm scenarioTemplateForm, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScenarioTemplates], ct))
+                throw new ForbiddenException();
+
             var createdScenarioTemplate = await _scenarioTemplateService.CreateAsync(scenarioTemplateForm, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdScenarioTemplate.Id }, createdScenarioTemplate);
         }
@@ -139,6 +109,10 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "copyScenarioTemplate")]
         public async STT.Task<IActionResult> Copy(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScenarioTemplates], ct)
+                || !await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ViewScenarioTemplates], [ScenarioTemplatePermission.ViewScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             var newScenarioTemplate = await _scenarioTemplateService.CopyAsync(id, ct);
             return CreatedAtAction(nameof(this.Get), new { id = newScenarioTemplate.Id }, newScenarioTemplate);
         }
@@ -159,6 +133,9 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "updateScenarioTemplate")]
         public async STT.Task<IActionResult> Update([FromRoute] Guid id, [FromBody] SAVM.ScenarioTemplateForm scenarioTemplateForm, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.EditScenarioTemplates], [ScenarioTemplatePermission.EditScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             var updatedScenarioTemplate = await _scenarioTemplateService.UpdateAsync(id, scenarioTemplateForm, ct);
             return Ok(updatedScenarioTemplate);
         }
@@ -178,6 +155,9 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "deleteScenarioTemplate")]
         public async STT.Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ManageScenarioTemplates], [ScenarioTemplatePermission.ManageScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             await _scenarioTemplateService.DeleteAsync(id, ct);
             return NoContent();
         }
